MNT: refactor DirectoryContainer
