+++ tests/test_protocols.py
@@ -12,7 +12,6 @@ from pytest_mock import MockFixture
 from tiled.access_control.access_policies import ALL_ACCESS
 from tiled.access_control.protocols import AccessPolicy
 from tiled.access_control.scopes import ALL_SCOPES
-from tiled.adapters.awkward_directory_container import DirectoryContainer
 from tiled.adapters.protocols import (
     ArrayAdapter,
     AwkwardAdapter,
@@ -22,7 +21,7 @@ from tiled.adapters.protocols import (
 )
 from tiled.ndslice import NDSlice
 from tiled.server.schemas import Principal, PrincipalType
-from tiled.storage import Storage
+from tiled.storage import DirectoryContainer, Storage
 from tiled.structures.array import ArrayStructure, BuiltinDtype
 from tiled.structures.awkward import AwkwardStructure
 from tiled.structures.core import Spec, StructureFamily
